Remove buffering for the delegate and add for Client Updates
Since we cannot ignore all the previous state changes made by de background URL session, we need to remove the buffering policy for the delegate.
While the UI can benefit from the buffering policy made. Since it is only interested into the latest change, but for the delegate thats not the case (there we want to capture all the changes).

For example we have a download A and B, t which represents Time, and we have the buffering policy of newest(1).
Now let's say we are at t = 2. Download A and B is progressing. Now we are at t = 3 and Download A is finished while Download B is still progressing. Now we are at t = 4 Download B is still progressing (since Download A is finished we do not get notified about its state anymore).
Now given the buffering policy is set to newest(1), we only capture the latest state inside the AsyncStream. Let's say we captured at t = 2 and because of a long processing task we then jump to t = 4. We would have missed the part where Download A finished and therefore would have not been able to notify the store about this change.

Without this buffering policy set to newest(1) we get all the buffered state changes. Meaning from t = 2 till t = 4 we get everything.